requirements.txt
Flask==2.3.2
numpy==1.26.0
opencv-python==4.8.0.74
scipy==1.11.3
Pillow==10.1.0
matplotlib==3.8.1
